Ingresso

+++ tcs/src/main/java/go/party/tcs/repository/FollowerRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import go.party.tcs.model.Follower;
 import go.party.tcs.model.Usuario;
 
-public interface FollowerRepository extends JpaRepository<Follower, Integer> {
+public interface FollowerRepository extends JpaRepository<Follower, Long> {
 
     List<Follower> findByFollowing(Usuario following);
     List<Follower> findByFollower(Usuario follower);
